Some Updated Files

+++ Cryptosystem/src/keyGen/KeyTuple.java
@@ -1,30 +1,32 @@
 package keyGen;
 
+import java.math.BigInteger;
+
 public class KeyTuple {
 
-	private long[][] privateKey;
-	private long[][] publicKey;
-	
-	public KeyTuple(long[][] pub, long[][] priv){
-		
-		publicKey = pub;
-		privateKey = priv;
-		
-	}
-	
-	public long[][] getPrivateKey() {
-		return privateKey;
-	}
-	public void setPrivateKey(long[][] privateKey) {
-		this.privateKey = privateKey;
-	}
-	public long[][] getPublicKey() {
-		return publicKey;
-	}
-	public void setPublicKey(long[][] publicKey) {
-		this.publicKey = publicKey;
-	}
-	
-	
-	
+        private BigInteger[][] privateKey;
+        private BigInteger[][] publicKey;
+        
+        public KeyTuple(BigInteger[][] pub, BigInteger[][] priv){
+                
+                publicKey = pub;
+                privateKey = priv;
+                
+        }
+        
+        public BigInteger[][] getPrivateKey() {
+                return privateKey;
+        }
+        public void setPrivateKey(BigInteger[][] privateKey) {
+                this.privateKey = privateKey;
+        }
+        public BigInteger[][] getPublicKey() {
+                return publicKey;
+        }
+        public void setPublicKey(BigInteger[][] publicKey) {
+                this.publicKey = publicKey;
+        }
+        
+        
+        
 }
